requirements.txt
python-telegram-bot==13.15
yt-dlp==2024.04.09